Move function sanitize_filename to createtorrentdialog.py
