+++ src/tribler/gui/dialogs/tests/test_createtorrentdialog.py
@@ -1,6 +1,6 @@
 from PyQt5.QtGui import QValidator
 
-from tribler.gui.dialogs.createtorrentdialog import TorrentNameValidator
+from tribler.gui.dialogs.createtorrentdialog import sanitize_filename, TorrentNameValidator
 
 
 def test_torrent_name_validator():
@@ -23,3 +23,9 @@ def test_torrent_name_validator():
 
     fixed_name = validator.fixup(invalid_name)
     assert_text_is_valid(validator, fixed_name, expected_to_be_valid=True)
+
+
+def test_sanitize_filename():
+    original_filename = "This \nIs \r\nA \tTorrent Name.torrent"
+    expected_sanitized_filename = "This Is A Torrent Name.torrent"
+    assert sanitize_filename(original_filename) == expected_sanitized_filename
